Move 'Change' from corro-api-types to corro-types
This struct is only required in Corrosion itself and not its consumers.
Exposing it via 'corro-api-types' is problematic, because it uses
unstable Rust feature that has recently been changed and, thus, requires
consumers to be on the same Rust version.
