+++ crates/corro-api-types/Cargo.toml
@@ -19,4 +19,3 @@ speedy = { workspace = true }
 strum = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
-corro-base-types = { path = "../corro-base-types" }
