+++ crates/corro-types/src/updates.rs
@@ -1,11 +1,12 @@
 use crate::actor::ActorId;
 use crate::agent::SplitPool;
+use crate::change::Change;
 use crate::pubsub::{unpack_columns, MatchCandidates, MatchableChange, MatcherError};
 use crate::schema::Schema;
 use antithesis_sdk::assert_sometimes;
 use async_trait::async_trait;
 use corro_api_types::sqlite::ChangeType;
-use corro_api_types::{Change, ColumnName, NotifyEvent, SqliteValueRef, TableName};
+use corro_api_types::{ColumnName, NotifyEvent, SqliteValueRef, TableName};
 use corro_base_types::CrsqlDbVersion;
 use indexmap::{map::Entry, IndexMap};
 use metrics::{counter, histogram, Counter};
